fix: refactor to follow SOLID principles
