+++ frontend/carta.html
@@ -138,7 +138,7 @@
           <hr />
           <p class="modal-price"></p>
           <p class="modal-text">Ingrese la cantidad a comprar.</p>
-          <input class="modal-input" type="number" value="1" min="0" />
+          <input class="modal-input" type="number" value="1" min="1" />
           <h3 class="modal-total"></h3>
           <p class="modal-textConfirm">¿Desea realizar la compra?</p>
         </div>
@@ -177,6 +177,6 @@
       integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
       crossorigin="anonymous"
     ></script>
-    <script src="./CartaScript.js" type="module"></script>
+    <script src="./js/carta.js" type="module"></script>
   </body>
 </html>
